Began customizing color.
I want to change it a bit more. Add colors representative of the other Atom notification errors, as well as an Atom-ish logo to the body. But that will be a later goal.

I'm also not super sold on how the colors are blending together, yet. I'll play with it a bit more.